fix test

+++ vlib/compiler/tests/typeof_test.v
@@ -4,7 +4,7 @@ fn test_typeof_on_simple_expressions() {
 	assert typeof(42) == 'int'
 	assert typeof(3.14) == 'f64'
 	assert typeof(2+2*10) == 'int'
-	assert typeof(1.0 * 12.2) == 'f32'
+	assert typeof(1.0 * 12.2) == 'f64'
 	assert typeof(a) == 'int'
 }
